# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Banamex CX MVP",
    "version": "0.1.0"
  },
  "paths": {
...
    "/upload": {
      "post": {
        "tags": [
          "upload"
        ],
        "summary": "Upload",
        "description": "Sube un `.txt` y lo procesa (parse → dedup → clasifica).\n\nDevuelve `{file_id, validation_summary, already_processed}` (`01 §8`).",
        "operationId": "upload_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/upload/{file_id}/status": {
      "get": {
        "tags": [
          "upload"
        ],
        "summary": "Upload Status",
        "description": "Devuelve el estado de procesamiento. En el flujo síncrono actual: `done`.",
        "operationId": "upload_status_upload__file_id__status_get",
        "parameters": [
          {
            "name": "file_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "File Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_upload_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UploadResponse": {
        "properties": {
          "file_id": {
            "type": "integer",
            "title": "File Id"
          },
          "validation_summary": {
            "$ref": "#/components/schemas/ValidationSummary"
          },
          "already_processed": {
            "type": "boolean",
            "title": "Already Processed"
          }
        },
        "type": "object",
        "required": [
          "file_id",
          "validation_summary",
          "already_processed"
        ],
        "title": "UploadResponse"
      },
      "UploadStatusResponse": {
        "properties": {
          "file_id": {
            "type": "integer",
            "title": "File Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "parsing",
              "classifying",
              "done",
              "error"
            ],
            "title": "Status"
          },
          "progress": {
            "type": "number",
            "title": "Progress"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "file_id",
          "status",
          "progress"
        ],
        "title": "UploadStatusResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "ValidationSummary": {
        "properties": {
          "files_processed": {
            "type": "integer",
            "title": "Files Processed"
          },
          "rows_loaded": {
            "type": "integer",
            "title": "Rows Loaded"
          },
          "rows_new": {
            "type": "integer",
            "title": "Rows New"
          },
          "rows_duplicated_ignored": {
            "type": "integer",
            "title": "Rows Duplicated Ignored"
          },
          "branches_detected": {
            "type": "integer",
            "title": "Branches Detected"
          },
          "period_available": {
            "prefixItems": [
              {
                "type": "string"
              },
              {
                "type": "string"
              }
            ],
            "type": "array",
            "maxItems": 2,
            "minItems": 2,
            "title": "Period Available"
          },
          "months_available": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Months Available"
          },
          "columns_detected": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Columns Detected"
          },
          "rows_valid": {
            "type": "integer",
            "title": "Rows Valid"
          },
          "rows_empty_verbatim": {
            "type": "integer",
            "title": "Rows Empty Verbatim"
          },
          "rows_invalid_nps": {
            "type": "integer",
            "title": "Rows Invalid Nps"
          },
          "rows_missing_branch": {
            "type": "integer",
            "title": "Rows Missing Branch"
          },
          "rows_duplicate_record_id": {
            "type": "integer",
            "title": "Rows Duplicate Record Id"
          },
          "rows_invalid_date": {
            "type": "integer",
            "title": "Rows Invalid Date"
          }
        },
        "type": "object",
        "required": [
          "files_processed",
          "rows_loaded",
          "rows_new",
          "rows_duplicated_ignored",
          "branches_detected",
          "period_available",
          "months_available",
          "columns_detected",
          "rows_valid",
          "rows_empty_verbatim",
          "rows_invalid_nps",
          "rows_missing_branch",
          "rows_duplicate_record_id",
          "rows_invalid_date"
        ],
        "title": "ValidationSummary"
      },
...
    }
  }
}